Move question answer mapping to new file
